requirements.txt
aiohttp==3.7.4
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
discord.py==@ git+https://github.com/Rapptz/discord.py@f6a74f74a7aed0879fc086805eae8873e745d0ea
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
yarl==1.7.2